Daniel

See merge request compsci308_2019spring/voogasalad_crackingopen!17

+++ src/Engine/src/Physics/CollisionHandler.java
@@ -31,14 +31,14 @@ public class CollisionHandler {
     }
 
     private boolean collides(GameObject obj1, GameObject obj2){
-        int height1 = obj1.getHeight();
-        int height2 = obj2.getHeight();
-        int width1 = obj1.getWidth();
-        int width2 = obj2.getWidth();
-        int x1 = obj1.getX();
-        int x2 = obj2.getX();
-        int y1 = obj1.getY();
-        int y2 = obj1.getY();
+        double height1 = obj1.getHeight();
+        double height2 = obj2.getHeight();
+        double width1 = obj1.getWidth();
+        double width2 = obj2.getWidth();
+        double x1 = obj1.getX();
+        double x2 = obj2.getX();
+        double y1 = obj1.getY();
+        double y2 = obj1.getY();
 
         if(x1 + width1 >= x2 || y1 + height1>= y2 || x2 + width2 >= x1 || y2 + height2>= y1){
             return true;
